requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
boto3~=1.34

dj-database-url==2.1.0
Django~=5.0
django-cors-headers==4.3.1
django-filter~=23.5
django-log-request-id==2.1.0
django-model-utils~=4.4
django-request-logging==0.7.5
django-storages~=1.14
djangorestframework==3.14.0
djangorestframework-simplejwt~=5.3
faiss-cpu==1.8.0
gunicorn>=21.2,<21.3
keras~=3.0.5
opencv-python>=4.1.1.26, !=4.9.0.80
pillow~=10.2.0
psycopg2==2.9.9
requests==2.31.0
tensorflow~=2.16.1

torch~=2.2.0
whitenoise==6.6.0
yolov7-package==0.0.12
